Fix exported types
