--- react/ExtensionPoint.tsx
@@ -1,3 +1,0 @@
-import { ExtensionPoint } from 'vtex.render-runtime'
-
-export default ExtensionPoint
