+++ react/components/Block.tsx
@@ -1,0 +1,3 @@
+import ExtensionPoint from './ExtensionPoint/index'
+
+export default ExtensionPoint
