+++ react/components/ExtensionPoint.tsx
@@ -1,0 +1,3 @@
+import ExtensionPoint from './ExtensionPoint/index'
+
+export default ExtensionPoint
