+++ react/components/Helmet.tsx
@@ -1,0 +1,3 @@
+import { Helmet } from 'react-helmet'
+
+export default Helmet
