+++ react/components/canUseDOM.ts
@@ -1,0 +1,1 @@
+export { canUseDOM } from 'exenv'
