+++ react/package.json
@@ -41,7 +41,7 @@
     "@types/debounce": "^1.0.0",
     "@types/eventemitter3": "^2.0.2",
     "@types/graphql": "^0.13.3",
-    "@types/history": "^4.6.2",
+    "@types/history": "^4.7.2",
     "@types/intl": "^1.2.0",
     "@types/jest": "^24.0.0",
     "@types/js-base64": "^2.3.1",
@@ -79,7 +79,7 @@
     "react-no-ssr": "^1.1.0",
     "regenerator-runtime": "^0.13.1",
     "ts-jest": "^24.0.1",
-    "typescript": "3.8.3"
+    "typescript": "3.9.7"
   },
   "husky": {
     "hooks": {
