+++ react/utils/messages.ts
@@ -1,4 +1,5 @@
 import { pluck, zipObj } from 'ramda'
+import { KeyedString } from '../typings/global'
 
 export const parseMessages = (messages: KeyedString[] | null) => {
   return messages && zipObj(pluck('key', messages), pluck('message', messages))
