+++ react/utils/routes.ts
@@ -9,6 +9,23 @@ import { fetchWithRetry } from './fetch'
 import { parseMessages } from './messages'
 import { isEnabled } from './flags'
 import { DeviceInfo } from './withDevice'
+import {
+  ApolloClientType,
+  PageQueryResponse,
+  FetchNavigationDataInput,
+  ParsedPageQueryResponse,
+  GraphQLResult,
+  FetchDefaultPages,
+  DefaultPagesQueryResponse,
+  ParsedDefaultPagesQueryResponse,
+  DefaultPagesQueryResult,
+  KeyedString,
+  PrefetchRouteData,
+  ContentResponse,
+  ServerPageResponse,
+  ParsedServerPageResponse,
+} from '../typings/global'
+import { RenderRuntime, Pages } from '../typings/runtime'
 
 const parsePageQueryResponse = (
   page: PageQueryResponse
