+++ types/index.d.ts
@@ -1,0 +1,13 @@
+declare module 'vtex.render-runtime' {
+  export interface RenderRuntime {
+    foo: 'bar'
+  }
+
+  export function useRuntime(): RenderRuntime
+
+  interface ChildBlock {
+
+  }
+
+  export function useChildBlock(): ChildBlock
+}
