+++ types/package.json
@@ -1,0 +1,13 @@
+{
+  "name": "@vtex/render-runtime-types",
+  "version": "8.0.0",
+  "description": "TypeScript definitions for Render Runtime",
+  "license": "MIT",
+  "main": "",
+  "types": "index.d.ts",
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/vtex-apps/render-runtime.git",
+    "directory": "types"
+  }
+}
